+++ services/madoc-ts/src/frontend/admin/pages/content/collections/collection-search-index.tsx
@@ -10,6 +10,15 @@ export const CollectionSearchIndex = () => {
 
   return (
     <div>
+      <p>Index manifests in the collection search index to:</p>
+      <ul>
+        <li>Add all manifests in this collection to the collection search</li>
+        <li> and/or delete manifests that have been deleted from this collection from the collection search.</li>
+      </ul>
+      <p>
+        This process may take some time, and will be run in the background. Once pressed you can navigate away from this
+        page.
+      </p>
       <Button disabled={isLoading} onClick={() => indexContext()}>
         Index all manifests {isLoading ? ` ${percent}%` : null}
       </Button>