+++ services/madoc-ts/src/frontend/admin/pages/enrichment/search-indexing.tsx
@@ -8,6 +8,7 @@ import { FacetConfig, MetadataFacetEditor } from '../../../shared/features/Metad
 import { useApi } from '../../../shared/hooks/use-api';
 import { apiHooks } from '../../../shared/hooks/use-api-query';
 import { AdminHeader } from '../../molecules/AdminHeader';
+import { HelpMessage } from '../../../shared/callouts/HelpMessage';
 
 export const SearchIndexingPage: React.FC = () => {
   const currentConfig = apiHooks.getSiteSearchFacetConfiguration(() => [], {
@@ -36,16 +37,36 @@ export const SearchIndexingPage: React.FC = () => {
           { label: 'Site admin', link: '/' },
           { label: 'Search indexing', active: true, link: `/enrichment/search-indexing` },
         ]}
-        title="Search indexing"
-        subtitle={
-          <>
-            <Button disabled={reindexStatus.isLoading} onClick={() => reindexFullSize()}>
-              {t('Reindex full site')}
-            </Button>
-          </>
-        }
+        title="Site search indexing"
+        subtitle="Site-wide search metadata and Site search indexing"
       />
       <WidePage>
+        <HelpMessage $margin title="About site indexing">
+          <p>
+            Decide which manifest metadata fields and its values are shown in the general site search, project searches
+            and collection searches. You do this by dragging metadata fields from the list on the right. You can also
+            combine multiple fields into one field, and rename if necessary. Once configured, up to 100 results will be
+            shown for each metadata item on the site search page.
+          </p>
+          <p>
+            When nothing is added from the list on the right, all metadata fields will be shown in the general site
+            search, project searches and collection searches. The number of items will be limited to 10 most common
+            metadata values.
+          </p>
+          <p>
+            If you are seeing missing Manifests or Collections in the search, you can trigger a full reindex of the
+            site. This will take some time, and will be run in the background. Once pressed you can navigate away from
+            this page.
+            <Button
+              $primary
+              disabled={reindexStatus.isLoading}
+              onClick={() => reindexFullSize()}
+              style={{ marginTop: 20 }}
+            >
+              {t('Reindex all manifests')}
+            </Button>
+          </p>
+        </HelpMessage>
         {currentConfig.data ? (
           <MetadataFacetEditor
             facets={currentConfig.data.facets}